Move RegionSelector component to /wca
